Added Prints to Debug
Encountered an issue with the deployed app Added print statement to help with debugging in future. The error was not because of the code but probably related to streamlit. It is solved.Just added print statements in each function for future debugging as they will be in CLI not in the streamlit WebApp. So no issues.